labels

+++ src/com/JakeKrayger/nn/training/loss/BinCrossEntropy.java
@@ -3,17 +3,16 @@ package src.com.JakeKrayger.nn.training.loss;
 import org.ejml.simple.SimpleMatrix;
 
 public class BinCrossEntropy extends Loss {
-    public double execute(SimpleMatrix activations, int[] labels) {
+    public double execute(SimpleMatrix activations, double[] labels) {
         double sumLoss = 0.0;
         double n = activations.getNumElements();
-        
+
         for (int i = 0; i < n; i++) {
             double pred = activations.get(i);
-            int y = labels[i];
+            double y = labels[i];
             // need to prevent log(0)
             sumLoss += -(y * Math.log(pred) + (1 - y) * Math.log(1 - pred));
         }
-
         return sumLoss / n;
     }
 }
